requirements.txt
PyAudio
DateTime
setuptools
numpy
pydub
scipy
Wave
typing
psutil
PyYAML
RapidFuzz
pathlib
scikit-learn
fuzzywuzzy
vosk
noisereduce
pvporcupine